Rename 'cargo/audio-test' crate to 'cargo/jni'
