+++ cargo/jni/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "audio-test"
+name = "musicbx-jni"
 version = "0.1.0"
 edition = "2021"
