+++ cargo/jni/src/configuration.rs
@@ -1,50 +1,25 @@
-use std::ops::Deref;
-use std::sync::Arc;
-use std::thread::sleep;
-use std::time::Duration;
-
-use jni::objects::{JClass, JObject, JString, JValue};
+use jni::objects::{JClass, JObject, JValue};
 use jni::JNIEnv;
 use lazy_static::lazy_static;
-use serde::de::Unexpected::Str;
-use serde_json::json;
+use serde::{Deserialize, Serialize};
+use std::sync::Arc;
 
-use crate::configuration::{Configuration, OutputConfiguration, SampleRateConfiguration};
-use crate::model::Module;
-
-mod configuration;
-mod model;
-
-#[no_mangle]
-pub extern "system" fn Java_ru_pema4_musicbx_service_PlaybackService_start(
-    _env: JNIEnv,
-    _class: JClass,
-) {
-    println!("Backend started");
+#[derive(PartialEq, Debug, Serialize, Deserialize)]
+pub struct Configuration {
+    pub output: OutputConfiguration,
 }
 
-#[no_mangle]
-pub extern "system" fn Java_ru_pema4_musicbx_service_PlaybackService_addModule(
-    env: JNIEnv,
-    _class: JClass,
-    module_json: JString,
-) {
-    let module_json = env.get_string(module_json).unwrap();
-    let module_json_bytes: &[u8] = module_json.to_bytes();
-    let module: serde_json::Result<Module> = serde_json::from_slice(module_json_bytes);
-    println!("{:?}", module);
+#[derive(PartialEq, Debug, Serialize, Deserialize)]
+pub struct OutputConfiguration {
+    pub current: String,
+    pub available: Vec<String>,
+    pub sample_rate: SampleRateConfiguration,
 }
 
-#[no_mangle]
-pub extern "system" fn Java_ru_pema4_musicbx_service_PlaybackService_removeModule(
-    env: JNIEnv,
-    _class: JClass,
-    module_json: JString,
-) {
-    let module_json = env.get_string(module_json).unwrap();
-    let module_json_bytes: &[u8] = module_json.to_bytes();
-    let module: serde_json::Result<Module> = serde_json::from_slice(module_json_bytes);
-    println!("{:?}", module);
+#[derive(PartialEq, Debug, Serialize, Deserialize)]
+pub struct SampleRateConfiguration {
+    pub current: f64,
+    pub available: Vec<f64>,
 }
 
 #[no_mangle]
