+++ cargo/jni/src/model.rs
